refactor(utils): rewrite all imports and delete mtui/utils.py
Mechanical rewrite of every in-tree import of the form
`from mtui.utils import X` (or `..utils` / `.utils` variants) to
point at the new homes established by the preceding five extraction
commits. The mapping is deterministic:

  green / red / yellow / blue     -> mtui.colors
  complete_choices(_filelist)     -> mtui.completion
  chdir / ensure_dir_exists /
    atomic_write_file / timestamp -> mtui.fileops
  termsize / filter_ansi /
    prompt_user / page            -> mtui.term
  DictWithInjections / SUTParse /
    requires_update               -> mtui.misc

Mixed-symbol imports are split across multiple `from` lines (one per
destination module); ruff's isort then resorts the import block.
50 production and test files touched. `mtui/utils.py` is deleted in
the same commit; out-of-tree consumers importing from it will get an
ImportError on upgrade (documented in CHANGELOG).

tests/test_utils.py is the only remaining importer of mtui.utils; it
is rewritten and split into the new tests/test_{colors,fileops,misc}.py
files in the next commit.

Sixth commit of the Phase 5b / C8 split (PLAN.md item 11).
