+++ mtui/actions/install.py
@@ -3,7 +3,7 @@
 from string import Template
 
 from ..messages import MissingInstallerError
-from ..utils import DictWithInjections
+from ..misc import DictWithInjections
 
 #: A dictionary of command templates for installing packages using zypper.
 zypper_install = {"command": Template("zypper -n in -y -l $packages")}
