+++ mtui/commands/addhost.py
@@ -3,7 +3,7 @@
 import concurrent.futures
 
 from mtui.commands import Command
-from mtui.utils import complete_choices
+from mtui.completion import complete_choices
 
 
 class AddHost(Command):
