+++ mtui/commands/commit.py
@@ -5,7 +5,8 @@ from argparse import REMAINDER
 from logging import getLogger
 
 from mtui.commands import Command
-from mtui.utils import complete_choices, requires_update
+from mtui.completion import complete_choices
+from mtui.misc import requires_update
 
 logger = getLogger("mtui.command.commit")
