+++ mtui/commands/downgrade.py
@@ -5,8 +5,9 @@ from traceback import format_exc
 
 from mtui.argparse import ArgumentParser
 from mtui.commands import Command
+from mtui.completion import complete_choices
 from mtui.messages import NoRefhostsDefinedError
-from mtui.utils import complete_choices, requires_update
+from mtui.misc import requires_update
 
 logger = getLogger("mtui.command.downgrade")
