+++ mtui/commands/edit.py
@@ -6,7 +6,8 @@ from subprocess import check_call
 
 from mtui.argparse import ArgumentParser
 from mtui.commands import Command
-from mtui.utils import complete_choices_filelist, requires_update
+from mtui.completion import complete_choices_filelist
+from mtui.misc import requires_update
 
 logger = getLogger("mtui.command.edit")
