+++ mtui/commands/hostslock.py
@@ -3,7 +3,7 @@
 from argparse import REMAINDER
 
 from mtui.commands import Command
-from mtui.utils import complete_choices
+from mtui.completion import complete_choices
 
 
 class HostLock(Command):
