+++ mtui/commands/hoststate.py
@@ -3,8 +3,8 @@
 from logging import getLogger
 
 from mtui.commands import Command
+from mtui.completion import complete_choices
 from mtui.types import ExecutionMode
-from mtui.utils import complete_choices
 
 logger = getLogger("mtui.command.hoststate")
