+++ mtui/commands/loadtemplate.py
@@ -2,9 +2,10 @@
 
 from mtui.argparse import ArgumentParser
 
+from ..completion import complete_choices
 from ..messages import TestReportNotLoadedError
+from ..term import prompt_user
 from ..types.updateid import AutoOBSUpdateID, KernelOBSUpdateID
-from ..utils import complete_choices, prompt_user
 from . import Command
