+++ mtui/commands/quit.py
@@ -6,7 +6,7 @@ from contextlib import suppress
 
 from mtui.argparse import ArgumentParser
 from mtui.commands import Command
-from mtui.utils import complete_choices
+from mtui.completion import complete_choices
 
 
 class Quit(Command):
