+++ mtui/commands/reload.py
@@ -4,7 +4,7 @@ from logging import getLogger
 
 from mtui.argparse import ArgumentParser
 from mtui.commands import Command
-from mtui.utils import complete_choices
+from mtui.completion import complete_choices
 
 logger = getLogger("mtui.commands.reload")
