+++ mtui/commands/removehost.py
@@ -4,7 +4,7 @@ import concurrent.futures
 
 from mtui.argparse import ArgumentParser
 from mtui.commands import Command
-from mtui.utils import complete_choices
+from mtui.completion import complete_choices
 
 
 class RemoveHost(Command):
