+++ mtui/commands/setrepo.py
@@ -4,8 +4,9 @@ from logging import getLogger
 
 from mtui.argparse import ArgumentParser
 from mtui.commands import Command
+from mtui.completion import complete_choices
+from mtui.misc import requires_update
 from mtui.target.locks import LockedTargets, TargetLockedError
-from mtui.utils import complete_choices, requires_update
 
 logger = getLogger("mtui.command.setrepo")
