--- mtui/utils.py
@@ -1,22 +1,0 @@
-from .colors import blue, green, red, yellow  # noqa: F401  # re-exported for callers
-from .completion import (  # noqa: F401  # re-exported for callers
-    complete_choices,
-    complete_choices_filelist,
-)
-from .fileops import (  # noqa: F401  # re-exported for callers
-    atomic_write_file,
-    chdir,
-    ensure_dir_exists,
-    timestamp,
-)
-from .misc import (  # noqa: F401  # re-exported for callers
-    DictWithInjections,
-    SUTParse,
-    requires_update,
-)
-from .term import (  # noqa: F401  # re-exported for callers
-    filter_ansi,
-    page,
-    prompt_user,
-    termsize,
-)
